chore: cleanup sorting utils
